docker image파일 생성
