+++ web_crawler/output/failed_urls.txt
@@ -1,0 +1,55 @@
+https://www.jjan.kr/article/20250313580209
+https://www.nongmin.com/article/20250121500316
+https://news.mt.co.kr/mtview.php?no=2025012401031186639
+https://www.donga.com/news/Society/article/all/20241217/130662494/2
+https://www.edaily.co.kr/News/Read?newsId=03076646639117208&amp;mediaCodeNo=257&amp;OutLnkChk=Y
+https://www.edaily.co.kr/News/Read?newsId=05205366639085392&mediaCodeNo=257&OutLnkChk=Y
+https://www.boannews.com/media/view.asp?idx=134173&kind=
+https://www.hankyung.com/article/2024102581287
+https://www.mk.co.kr/news/society/11120483
+https://www.etoday.co.kr/news/view/2393116
+http://news.moneytoday.co.kr/view/mtview.php?no=2024081416463842438
+https://www.etoday.co.kr/news/view/126518
+https://science.ytn.co.kr/program/view.php?mcd=0082&hcd=&key=202408021627457166
+https://www.sedaily.com/NewsView/2DBTSA778Z
+https://www.newsis.com/view/NISX20240802_0002835991
+http://www.segye.com/content/html/2024/05/01/20240501505262.html
+https://www.nocutnews.co.kr/news/6060582
+https://www.kihoilbo.co.kr/news/articleView.html?idxno=1088782
+https://www.boannews.com/media/view.asp?idx=129836&kind=
+https://www.fnnews.com/news/202305090950301041
+https://www.fnnews.com/news/202307181424576426
+https://newsis.com/view/?id=NISX20230713_0002374963&cID=15001&pID=15000
+https://www.mk.co.kr/news/society/10778429
+https://newsis.com/view/?id=NISX20230615_0002340652&cID=10201&pID=10200
+http://www.sentv.co.kr/news/view/655685
+https://www.newsquest.co.kr/news/articleView.html?idxno=100337
+https://www.mbn.co.kr/news/economy/4823050
+https://www.jjan.kr/article/20250313580209
+https://www.munhwa.com/news/view.html?no=2025011901039905098001
+https://www.news1.kr/local/jeonbuk/5669701
+https://www.moneys.co.kr/article/2024112913000071858
+https://www.donga.com/news/Society/article/all/20241217/130662494/2
+https://www.edaily.co.kr/News/Read?newsId=05205366639085392&mediaCodeNo=257&OutLnkChk=Y
+https://www.pinpointnews.co.kr/news/articleView.html?idxno=299943
+https://www.chungnamilbo.co.kr/news/articleView.html?idxno=790972
+https://www.mk.co.kr/news/society/11120483
+https://www.hankyung.com/article/2024081287001
+https://www.etoday.co.kr/news/view/2393116
+https://www.etoday.co.kr/news/view/126518
+https://science.ytn.co.kr/program/view.php?mcd=0082&hcd=&key=202408021627457166
+https://www.sedaily.com/NewsView/2DBTSA778Z
+http://www.boannews.com/media/view.asp?idx=130683&kind=&sub_kind=
+http://www.segye.com/content/html/2024/05/01/20240501505262.html
+https://biz.sbs.co.kr/article/20000178390?division=NAVER
+https://www.nocutnews.co.kr/news/6060582
+https://www.kihoilbo.co.kr/news/articleView.html?idxno=1088782
+https://www.boho.or.kr/kr/bbs/view.do?searchCnd=&bbsId=B0000133&searchWrd=&menuNo=205020&pageIndex=1&categoryCode=&nttId=71132
+https://www.fnnews.com/news/202305090950301041
+https://www.fnnews.com/news/202307181424576426
+https://newsis.com/view/?id=NISX20230713_0002374963&cID=15001&pID=15000
+https://news.mt.co.kr/mtview.php?no=2023071314551957393
+https://www.donga.com/news/Economy/article/all/20230713/120221797/1
+https://newsis.com/view/?id=NISX20230615_0002340652&cID=10201&pID=10200
+http://www.sentv.co.kr/news/view/655685
+https://www.mbn.co.kr/news/economy/4823050
