+++ web_crawler/requirements.txt
@@ -1,0 +1,105 @@
+aiohappyeyeballs==2.6.1
+aiosignal==1.3.2
+alembic==1.15.2
+annotated-types==0.7.0
+anyio==4.9.0
+async-timeout==4.0.3
+attrs==25.3.0
+babel==2.17.0
+beautifulsoup4==4.13.3
+certifi==2025.4.26
+cffi==1.17.1
+charset-normalizer==3.4.2
+click==8.1.8
+colorama==0.4.6
+courlan==1.3.2
+dataclasses-json==0.6.7
+dateparser==1.2.1
+distro==1.9.0
+docx2txt==0.9
+exceptiongroup==1.2.2
+fastapi==0.115.12
+frozenlist==1.5.0
+greenlet==3.1.1
+h11==0.14.0
+htmldate==1.9.3
+httpcore==1.0.7
+httptools==0.6.4
+httpx==0.28.1
+httpx-sse==0.4.0
+idna==3.10
+iniconfig==2.1.0
+jiter==0.9.0
+jsonpatch==1.33
+jsonpointer==3.0.0
+jusText==3.0.2
+langchainhub==0.1.21
+loguru==0.7.3
+lxml==5.3.2
+lxml_html_clean==0.4.2
+Mako==1.3.10
+MarkupSafe==3.0.2
+marshmallow==3.26.1
+multidict==6.3.2
+mypy-extensions==1.0.0
+numpy==2.2.4
+openai==1.70.0
+orjson==3.10.16
+outcome==1.3.0.post0
+packaging==24.2
+pandas==2.2.3
+pinecone==6.0.2
+pinecone-client==6.0.0
+pinecone-plugin-interface==0.0.7
+pluggy==1.5.0
+propcache==0.3.1
+pycparser==2.22
+pydantic==2.11.3
+pydantic-settings==2.8.1
+pydantic_core==2.33.1
+PySocks==1.7.1
+pytest==8.3.5
+pytest-asyncio==0.26.0
+pytest-socket==0.7.0
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.0
+pytz==2025.2
+PyYAML==6.0.2
+regex==2024.11.6
+requests==2.32.3
+requests-toolbelt==1.0.0
+selenium==4.33.0
+six==1.17.0
+sniffio==1.3.1
+sortedcontainers==2.4.0
+soupsieve==2.6
+SQLAlchemy==2.0.40
+starlette==0.46.2
+syrupy==4.9.1
+tenacity==9.1.2
+tiktoken==0.9.0
+tld==0.13
+tomli==2.2.1
+tqdm==4.67.1
+trafilatura==2.0.0
+trio==0.30.0
+trio-websocket==0.12.2
+types-requests==2.32.0.20250328
+typing-inspect==0.9.0
+typing-inspection==0.4.0
+typing_extensions==4.13.2
+tzdata==2025.2
+tzlocal==5.3.1
+ujson==5.10.0
+urllib3==2.4.0
+uvicorn==0.34.1
+watchfiles==1.0.5
+websocket-client==1.8.0
+websockets==15.0.1
+win32_setctime==1.2.0
+wincertstore==0.2
+wrapt==1.17.2
+wsproto==1.2.0
+yarl==1.19.0
+zipp==3.21.0
+zstandard==0.23.0
